Move emitting cfgs to pyo3-build-config
